backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
langchain==0.0.78
flask-cors==4.0.0  # or latest compatible version
requests==2.28.2
python-dotenv==1.0.0
azure-identity
azure-keyvault-secrets
quart==0.19.6
